Allow pickling of Track instances, closes #68
The Track class is incompatible with the pickle module because of the
magic __getattribute__ override, resulting in  "TypeError: 'NoneType'
object is not callable" error when unpickling (in Python 3 for example,
pickle tries to get the object's __reduce_ex__() function and then call
it).

This commit fixes it by implementing trivial __getstate__ and
__setstate__ pickle protocol support that simply uses __dict__ for the
state.